add undo and redo functionality
row state is stored in a stack.
changes are pushed onto the stack,
the last change is popped off when
the user does the undo action (Ctrl + Z).

redo state is also stored in a stack.
the row that gets popped off the row stack
gets pushed onto the redo stack.
the redo stack gets cleared when a new
update is pushed onto the row stack