transform creatures dataset
i wanted to get rid of fields that can be
easily derived, like STR_mod, to make the JSON smaller.
i also wanted more structured data to make it easier
to build features on top of it later on.
this commit creates the transform script and
updates the dataset with the transformed result