initial MobileInitiative component
i thought about how to adapt the existing
desktop Initiative interface to work on mobile.
i came to the conclusion that the spreadsheet
UI wouldn't work for mobile, so i went with a
card-based approach.

it's not complete, the cards aren't editable,
but it's an okay starting point for the layout.
not sure about the add button color